% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_fn_ground_temperature}
\alias{get_fn_ground_temperature}
\title{get_fn_ground_temperature}
\usage{
get_fn_ground_temperature(temp_ref, dtemp_ref, time_ref, diffusivity = 1e-06)
}
\arguments{
\item{temp_ref}{numeric, mean air temperature for the year (°C)}

\item{dtemp_ref}{numeric, amplitude for mean daily air temperature (°C)}

\item{time_ref}{numeric,  number of days after solstice when min temperature (days)}

\item{diffusivuity}{(m^2/s) thermal diffusivity (default 1.e-6, P. Begeury suggestion)}
}
\value{
function for ground-temperature
}
\description{
Generate a function to model the ground temperature
}
\details{
The function will have arguements: instant, depth (m)

The function will return an estimated ground-temperature
for the given instant(s) and depth

http://soilphysics.okstate.edu/software/SoilTemperature/document.pdf
}

