% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ground_temp_coef}
\alias{ground_temp_coef}
\title{ground_temp_coef}
\usage{
ground_temp_coef(instant, temp, n_hemi = TRUE)
}
\arguments{
\item{instant}{POSIXct instants}

\item{temp}{numeric ambient air (dry-bulb) temperature}

\item{n_hemi}{logical, indicates fit for northern hemisphere}
}
\value{
list of coefficients
}
\description{
Given a vector on instants and a vector of air-temperautures, return a list of
coefficients for the surface temperature.
}
\details{
A sinusoid is fit to the time/tempurature pairs.
The coefficients of the sine wave (temp_ref, dtemp_ref, time_ref) are returned.

http://soilphysics.okstate.edu/software/SoilTemperature/document.pdf
}

