% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{tropical_year}
\alias{tropical_year}
\title{Tropical year}
\format{A list with two members:
\describe{
  \item{time_ref}{POSIXct, instant describing December 2010 solstice}
  \item{dtime_ref}{numeric, length of a tropical year, in seconds}
}}
\source{
\url{http://en.wikipedia.org/wiki/Tropical_year}

\url{http://en.wikipedia.org/wiki/December_solstice}
}
\usage{
tropical_year
}
\description{
Description of a tropical year.
}
\keyword{datasets}

