% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{wx_2011}
\alias{wx_2011}
\title{wx_2011}
\format{A data frame of 8760 observations of 5 variables:
\describe{
  \item{instant}{POSIXct, time of the observation}
  \item{tmp_db}{numeric, dry-bulb temperature (°C)}
  \item{tmp_wb}{numeric, wet-bulb temperature, (°C)}
  \item{rel_humid}{numeric, relative humidity (1)}
  \iten{p_sl}{numeric, sea-level pressure (Pa)}
}}
\usage{
wx_2011
}
\description{
Weather dataset for 2011
}
\keyword{datasets}

